moved testAccCheckDatadogApplicationKeyExists to data application key test
