+++ datadog/tests/resource_datadog_application_key_test.go
@@ -180,34 +180,6 @@ resource "datadog_application_key" "foo" {
 }`, uniq, serviceAccount)
 }
 
-func testAccCheckDatadogApplicationKeyExists(accProvider func() (*schema.Provider, error), n string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		provider, _ := accProvider()
-		providerConf := provider.Meta().(*datadog.ProviderConfiguration)
-		apiInstances := providerConf.DatadogApiInstances
-		auth := providerConf.Auth
-
-		if err := datadogApplicationKeyExistsHelper(auth, s, apiInstances, n); err != nil {
-			return err
-		}
-		return nil
-	}
-}
-
-func datadogApplicationKeyExistsHelper(ctx context.Context, s *terraform.State, apiInstances *utils.ApiInstances, name string) error {
-	id := s.RootModule().Resources[name].Primary.ID
-	if v, ok := s.RootModule().Resources[name].Primary.Attributes["service_account"]; ok {
-		if _, _, err := apiInstances.GetServiceAccountsApiV2().GetServiceAccountApplicationKey(ctx, v, id); err != nil {
-			return fmt.Errorf("received an error retrieving service account application key %s", err)
-		}
-	} else {
-		if _, _, err := apiInstances.GetKeyManagementApiV2().GetCurrentUserApplicationKey(ctx, id); err != nil {
-			return fmt.Errorf("received an error retrieving application key %s", err)
-		}
-	}
-	return nil
-}
-
 func testAccCheckDatadogApplicationKeyValueMatches(accProvider func() (*schema.Provider, error), n string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		provider, _ := accProvider()
